+++ CHANGELOG.md
@@ -9,6 +9,20 @@ You may also reference live-docs based on the openapi specification here:
 https://tapis-project.github.io/live-docs
 
 ---------------------------------------------------------------------------
+## 1.9.0 - 2025-??-??
+
+Incremental improvements.
+
+### Breaking changes:
+- Remove support for deprecated runtime option SINGULARITY_START
+
+### New features:
+- Log service configuration parameters at start-up.
+
+### Bug fixes:
+- None
+
+---------------------------------------------------------------------------
 ## 1.8.2 - 2025-04-29
 
 Incremental improvements. Remove unused pretty-print option from resource classes.